+++ lib/qonfig/file_data_resolving.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+# @api private
+# @since 0.26.0
+module Qonfig::FileDataResolving
+  require_relative 'file_data_resolving/resolver'
+  require_relative 'file_data_resolving/mixin'
+end