+++ lib/qonfig/loaders/basic.rb
@@ -30,10 +30,11 @@ class Qonfig::Loaders::Basic
     #
     # @api private
     # @since 0.5.0
-    def load_file(file_path, fail_on_unexist: true)
-      load(::File.read(file_path))
-    rescue Errno::ENOENT => error
-      fail_on_unexist ? (raise Qonfig::FileNotFoundError, error.message) : load_empty_data
+    def load_file(file_path, fail_on_unexist: true, **options)
+      data = Qonfig::FileDataResolving::Resolver.resolve!(file_path, **options)
+      load(data)
+    rescue Qonfig::FileNotFoundError
+      fail_on_unexist ? raise : load_empty_data
     end
   end
 end