+++ lib/qonfig/plugins/vault/loaders/vault.rb
@@ -15,18 +15,19 @@ class Qonfig::Loaders::Vault < Qonfig::Loaders::Basic
   class << self
     # @param path [String, Pathname]
     # @option fail_on_unexist [Boolean]
+    # @option version [String, Integer]
     # @return [Object]
     #
     # @raise [Qonfig::FileNotFoundError]
     #
     # @api private
     # @since 0.25.0
-    def load_file(path, fail_on_unexist: true)
-      data = ::Vault.with_retries(Vault::HTTPError) do
-        ::Vault.logical.read(path.to_s)&.data&.dig(:data)
-      end
-      raise Qonfig::FileNotFoundError, "Path #{path} not exist" if data.nil? && fail_on_unexist
+    def load_file(path, fail_on_unexist: true, transform_values: true, version: nil, use_kv: true)
+      data = load_data(path, version, use_kv)
+      raise Qonfig::FileNotFoundError, "Path #{path} not exist" if data == nil && fail_on_unexist
       result = data || empty_data
+      return result unless transform_values
+
       deep_transform_values(result)
     rescue Vault::VaultError => error
       raise(Qonfig::VaultLoaderError.new(error.message).tap do |exception|
@@ -43,6 +44,25 @@ class Qonfig::Loaders::Vault < Qonfig::Loaders::Basic
     end
 
     private
+
+    # @param file_path [String]
+    # @param version [Integer]
+    # @return [Object]
+    #
+    # @api private
+    # @since 0.26.0
+    def load_data(file_path, version, use_kv)
+      response = ::Vault.with_retries(::Vault::HTTPError) do
+        if use_kv
+          mount_path, secret_path = file_path.to_s.split(::File::Separator, 2)
+          ::Vault.kv(mount_path).read(secret_path, version)
+        else
+          ::Vault.logical.read(file_path.to_s)
+        end
+      end
+
+      response&.data
+    end
 
     # @param vault_data [Hash<Object,Object>]
     # @return [Object]