+++ spec/features/plugins/vault/expose_vault_spec.rb
@@ -1,11 +1,9 @@
 # frozen_string_literal: true
 
+require_relative 'context'
+
 describe 'Plugins(vault): expose vault', plugin: :vault do
   before { stub_const('VaultConfig', vault_class) }
-
-  before { allow(Vault).to receive(:logical).and_return(logical_double) }
-
-  let(:logical_double) { instance_double(Vault::Logical) }
 
   let(:returned_data) do
     instance_double(Vault::Secret).tap do |instance|
@@ -13,17 +11,17 @@ describe 'Plugins(vault): expose vault', plugin: :vault do
     end
   end
   let(:secret_data) do
-    { data: { production: { kek: 'pek', cheburek: true }, other_key: '<%= 1 + 1 %>' } }
+    { production: { kek: 'pek', cheburek: true }, other_key: '<%= 1 + 1 %>' }
   end
 
   let(:vault_class) do
     Class.new(Qonfig::DataSet) do
       setting :based_on_path do
-        expose_vault 'kv/data/path_based', via: :path, env: :production
+        expose_vault 'kv/data/path_based', via: :path, env: :production, use_kv: false
       end
 
       setting :based_on_env_key do
-        expose_vault 'kv/data/env_key', via: :env_key, env: 'production'
+        expose_vault 'kv/data/env_key', via: :env_key, env: 'production', use_kv: false
       end
     end
   end
@@ -45,7 +43,7 @@ describe 'Plugins(vault): expose vault', plugin: :vault do
     specify 'raises an error' do
       expect do
         Class.new(Qonfig::DataSet) do
-          expose_vault 'kv/data/path_based', via: Object.new, env: :production
+          expose_vault 'kv/data/path_based', via: Object.new, env: :production, use_kv: false
         end
       end.to raise_error(Qonfig::ArgumentError)
     end
@@ -55,7 +53,7 @@ describe 'Plugins(vault): expose vault', plugin: :vault do
     specify 'raises an error' do
       expect do
         Class.new(Qonfig::DataSet) do
-          expose_vault 'kv/data/path_based', via: :path, env: Object.new
+          expose_vault 'kv/data/path_based', via: :path, env: Object.new, use_kv: false
         end
       end.to raise_error(Qonfig::ArgumentError)
     end
@@ -65,7 +63,7 @@ describe 'Plugins(vault): expose vault', plugin: :vault do
     specify 'raises an error' do
       expect do
         Class.new(Qonfig::DataSet) do
-          expose_vault 'kv/data/path_based', via: :kek, env: :production
+          expose_vault 'kv/data/path_based', via: :kek, env: :production, use_kv: false
         end
       end.to raise_error(Qonfig::ArgumentError)
     end
@@ -75,7 +73,7 @@ describe 'Plugins(vault): expose vault', plugin: :vault do
     specify 'raises an error' do
       expect do
         Class.new(Qonfig::DataSet) do
-          expose_vault 'kv/data/path_based', via: :path, env: ''
+          expose_vault 'kv/data/path_based', via: :path, env: '', use_kv: false
         end
       end.to raise_error(Qonfig::ArgumentError)
     end
@@ -84,7 +82,7 @@ describe 'Plugins(vault): expose vault', plugin: :vault do
   context "when provided key doesn't exist" do
     let(:vault_class) do
       Class.new(Qonfig::DataSet) do
-        expose_vault 'kv/data/env_key', via: :env_key, env: 'kekduction'
+        expose_vault 'kv/data/env_key', via: :env_key, env: 'kekduction', use_kv: false
       end
     end
 
@@ -99,7 +97,8 @@ describe 'Plugins(vault): expose vault', plugin: :vault do
     let(:vault_class) do
       Class.new(Qonfig::DataSet) do
         setting :based_on_env_key do
-          expose_vault 'kv/data/env_key', via: :env_key, env: 'production', strict: false
+          expose_vault 'kv/data/env_key',
+                       via: :env_key, env: 'production', strict: false, use_kv: false
         end
       end
     end